il mio pc si rifiuta

+++ alimentatore/alimentatore.c
@@ -1,10 +1,9 @@
 #include <time.h>
 #include <stdio.h>
+#include <errno.h>
 #include <fcntl.h>
 #include <stdlib.h>
 #include <signal.h>
-#include <sys/wait.h>
-#include <errno.h>
 
 #include "model.h"
 #include "lib/sem.h"
@@ -12,14 +11,14 @@
 #include "lib/fifo.h"
 #include "lib/shmem.h"
 
-int running();
-
-void signal_handler(int signum);
+int MEANINGFUL_SIGNALS[] = {SIGALRM, -1};
 
 struct Model *model = NULL;
-sig_atomic_t sig = -1;
+extern sig_atomic_t sig;
+
 
 int main(int argc, char *argv[]) {
+    print(D, "Alimentatore: %d\n", getpid());
     if (argc != 2) {
         print(E, "Usage: %s <shmid>\n", argv[0]);
         exit(EXIT_FAILURE);
@@ -55,15 +54,27 @@ int main(int argc, char *argv[]) {
     prargs("atomo", &argvc, &buf, 2, ITC_SIZE);
     sprintf(argvc[1], "%d", model->res->shmid);
 
+    struct sembuf sops[2];
+    sem_buf(&sops[0], SEM_INIBITORE_ON, 0, IPC_NOWAIT);
+    sem_buf(&sops[1], SEM_ALIMENTATORE, -1, 0);
+
     timer_t timer = timer_start(STEP_ALIMENTAZIONE);
     while (running()) {
-        for (int i = 0; sig != SIGTERM && i < N_NUOVI_ATOMI; i++) {
-            sprintf(argvc[2], "%d", rand_between(MIN_N_ATOMICO, N_ATOM_MAX));
-            pid_t child_pid = fork_execve(argvc);
-            if (child_pid != -1) {
-                fifo_add(model->res->fifo_fd, &child_pid, sizeof(pid_t));
-            } else {
-                kill(model->ipc->master, SIGMELT);
+        int n_atoms = 0;
+        while (sig != SIGTERM && n_atoms < N_NUOVI_ATOMI) {
+            if (sem_op(model->ipc->semid, sops, 2) == 0 || errno == EAGAIN) {
+                sprintf(argvc[2], "%d", rand_between(MIN_N_ATOMICO, N_ATOM_MAX));
+                pid_t child_pid = fork_execve(argvc);
+
+                if (child_pid != -1) {
+                    fifo_add(model->res->fifo_fd, &child_pid, sizeof(pid_t));
+                    n_atoms++;
+                } else {
+                    kill(model->ipc->master, SIGMELT);
+                    break;
+                }
+            } else if (errno == EINTR && sig == SIGALRM) {
+                break;
             }
         }
     }
@@ -88,30 +99,30 @@ void cleanup() {
 }
 
 
-int running() {
-    // while no meaningful signal is received
-    // - SIGTERM, means termination
-    // - SIGALRM, means STEP_ALIMENTAZIONE expired
-    while (sig != SIGTERM && sig != SIGALRM) {
-        // wait for children processes to terminate
-        while (wait(NULL) == -1) {
-            // if this process has no children
-            if (errno == ECHILD) {
-                // wait until a signal is received
-                pause();
-                // when pause is interrupted by a signal,
-                // break the inner loop so that meaningful
-                // signals are checked by the outer one
-                break;
-            }
-        }
-    }
-
-    int ret = sig != SIGTERM;
-    sig = -1;
-    return ret;
-}
-
-void signal_handler(int signum) {
-    sig = signum;
-}
+//int running() {
+//    // while no meaningful signal is received
+//    // - SIGTERM, means termination
+//    // - SIGALRM, means STEP_ALIMENTAZIONE expired
+//    while (sig != SIGTERM && sig != SIGALRM) {
+//        // wait for children processes to terminate
+//        while (wait(NULL) == -1) {
+//            // if this process has no children
+//            if (errno == ECHILD) {
+//                // wait until a signal is received
+//                pause();
+//                // when pause is interrupted by a signal,
+//                // break the inner loop so that meaningful
+//                // signals are checked by the outer one
+//                break;
+//            }
+//        }
+//    }
+//
+//    int ret = sig != SIGTERM;
+//    sig = -1;
+//    return ret;
+//}
+//
+//void signal_handler(int signum) {
+//    sig = signum;
+//}
